file(#30): ChatBox 파일명 변경
